+++ src/lib/Fullscreen.js
@@ -91,25 +91,32 @@ class Fullscreen extends EventEmitter {
      * Fires events when the fullscreen state changes
      *
      * @private
-     * @param {HTMLElement|Event} [el] - Fullscreen element
+     * @param {HTMLElement|Event} el - Fullscreen element
      * @return {void}
      */
     fullscreenchangeHandler = (el) => {
         let enter = false;
 
-        if (this.isSupported()) {
-            if (this.isFullscreen()) {
-                enter = true;
-            }
-        } else if (!this.isFullscreen(el)) {
-            enter = true;
-        }
+        enter = (this.isSupported() && this.isFullscreen()) || (!this.isSupported() && !this.isFullscreen(el));
 
         if (enter) {
+            this.focusFullscreenElement();
             this.emit('enter');
         } else {
             this.emit('exit');
         }
+    };
+
+    /**
+     * Focuses the element
+     *
+     * @private
+     * @return {void}
+     */
+    focusFullscreenElement = () => {
+        // Focus on the fullscreen element so keyboard
+        // events are triggered without an extra click
+        fscreen.fullscreenElement.focus();
     };
 
     /**